web-animations/animation-effect/animation-effect-composite.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationEffect composite attribute test</title>
<meta name="assert" content="The composite operation used to composite this animation effect with the animation stack, as specified by one of the CompositeOperation enumeration values">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationeffect-composite">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var effects = [newKeyframeEffect(), newMotionPathEffect()];
    effects.forEach(function(effect) {
        assert_equals(effect.composite, 'replace', 'Wrong default value of ' +
            'composite attribute of effect ' + type(effect));
    });
}, 'Test AnimationEffect.composite default value is \'replace\'');

test(function() {
    var effects = [newKeyframeEffect({composite: 'accumulate'}),
        newMotionPathEffect({composite: 'accumulate'})];
    effects.forEach(function(effect) {
        assert_equals(effect.composite, 'accumulate', type(effect) +
            '.composite should return the value specified in constructor arguments');
    });
}, 'Test AnimationEffect.composite attribute returns value set by constructor');

async_test(function(t) {
    var options = {composite: 'replace'};
    var effectsToTest = [[newKeyframeEffect(options), newKeyframeEffect(options)],
        [newMotionPathEffect(options), newMotionPathEffect(options)]];
    effectsToTest.forEach(function(effects) {
        var effect1 = effects[0];
        var effect2 = effects[1];
        var target = createDiv(t);

        var animation1 = new Animation(target, effect1, ANIMATION_END_TIME);
        var player1 = document.timeline.play(animation1);
        player1.currentTime = ANIMATION_END_TIME / 2 - 10;

        var animation2 = new Animation(target, effect2, ANIMATION_END_TIME);
        var player2 = document.timeline.play(animation2);
        player2.currentTime = 0;

        // player2 is created later than player1
        // so it has greater sequence number and higher priority
        setTimeout(t.step_func(function() {
            assert_approx_equals(window.getComputedStyle(target).top,
                getExpectedTop(100), EPSILON,
                type(effect2) + ' effect should replace the underlying value ' +
                    'in the animation stack');
        }), 110);
    });
}, 'Test AnimationEffect.composite: value \'replace\' forces the effect to replace ' +
    'an underlying value in the animation stack by its intermediate animation value');

async_test(function(t) {
    var options = {composite: 'add'};
    var effectsToTest = [[newKeyframeEffect(options), newKeyframeEffect(options)],
        [newMotionPathEffect(options), newMotionPathEffect(options)]];
    effectsToTest.forEach(function(effects) {
        var effect1 = effects[0];
        var effect2 = effects[1];
        var target = createDiv(t);

        var animation1 = new Animation(target, effect1, ANIMATION_END_TIME);
        var player1 = document.timeline.play(animation1);
        player1.currentTime = ANIMATION_END_TIME / 2 - 10;

        var animation2 = new Animation(target, effect2, ANIMATION_END_TIME);
        var player2 = document.timeline.play(animation2);
        player2.currentTime = ANIMATION_END_TIME / 2 - 10;

        setTimeout(t.step_func(function() {
            // Expected top property value is
            // 2 * (Animation top at the moment ANIMATION_END_TIME / 2 + 100)
            assert_approx_equals(window.getComputedStyle(target).top,
                2 * getExpectedTop(ANIMATION_END_TIME / 2 + 100), EPSILON,
                type(effect2) + 'effect should add its intermediate animation value ' +
                'to an underlying value in the animation stack');
        }), 110);
    });
}, 'Test AnimationEffect.composite: value \'add\' forces the effect to add ' +
    'its intermediate animation value to an underlying value in the animation stack');

// TODO add composite 'accumulate' test
</script>
</body>
